ENH: sixDoFRigidBodyDisplacement. rho field handling and new tutorial.
Removing rhoInf from uncoupledSixDoF... patch, not needed.  Adding
optional rhoName entry to sixDoF... so no rhoInf entry is needed for
simulations that have a rho field, and a rhoInf value is required and
read for those that don't.

New pimpleDyMFoam tutorial of fluttering wing, demonstrating the
sixDoFRigidBodyDisplacement restraints and constraints.