contactAngle: improved the handling of contact-angle in the VoF solvers.
Now there is an additional entry in the alphaContactAngle BCs to control the
gradient of alpha1 on the wall:

limit none;         // Calculate the gradient from the contact-angle without
                    // limiter
limit gradient;     // Limit the wall-gradient such that alpha1 remains bounded
                    // on the wall
limit alpha;        // Bound the calculated alpha1 on the wall
limit zeroGradient; // Set the gradient of alpha1 to 0 on the wall
                    // (i.e. previous behaviour)

Note that if any of the first three options are used the boundary condition on
p_rgh must set to guarantee that the flux is corrected to be zero at the wall
e.g.

    walls
    {
        type            fixedFluxPressure;
        adjoint         no;
    }

If "limit zeroGradient;" is used the pressure BCs can be left as before.